Refactored directory structure, switched to immutable map instead of list
