+++ src/controllers/index.js
@@ -1,2 +1,2 @@
 module.exports.Graph = require('./Graph.js');
-module.exports.InfluxDB = require('./InfluxDB.js');
+module.exports.InfluxDB = require('../influxdb/InfluxDB.js');
